+++ lib/src/format.dart
@@ -1,4 +1,5 @@
 import 'dart:math';
+import 'package:path/path.dart';
 
 /// Return number format including comma symbol.
 String numberFormat(dynamic number) {
@@ -10,6 +11,29 @@ String numberFormat(dynamic number) {
       RegExp(r'(\d)(?=(\d{3})+(?!\d))'), (Match match) => '${match[1]},');
 
   return formattedIntegerPart + decimalPart;
+}
+
+/// Extract the file name from the path. Include the extension if withExtension is `true`.
+String fileName(String filePath, [bool withExtension = false]) {
+  if (filePath.isEmpty) {
+    return '';
+  }
+
+  final String normalizedPath = filePath.replaceAll('\\', '/');
+
+  if (!normalizedPath.contains('/')) {
+    if (withExtension) {
+      return normalizedPath;
+    } else {
+      return basenameWithoutExtension(normalizedPath);
+    }
+  }
+
+  if (withExtension) {
+    return basename(normalizedPath);
+  } else {
+    return basenameWithoutExtension(normalizedPath);
+  }
 }
 
 /// Converts the file size in bytes to human-readable and returns it. The return value is a String and includes the file units (Bytes, MB, GB...). If the second optional argument value is included, you can display as many decimal places as you like.