+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `fileExt` method
 - Add `safeParseInt` method
 - Add `isEmail` method
+- Add `fileName` method
 
 ## 0.0.4 (2024-10-02)